requirements.txt
fastapi==0.136.1
langchain-oci==0.2.5
numpy==2.4.4
oci==2.173.1
oracleagentmemory==26.4.0
oracledb==3.4.2
python-dotenv==1.0.1
streamlit==1.57.0
uvicorn==0.46.0
